Align the 2026-09-04 seam catalog with shipped code
- Count channel-ingress as Slack, Discord, email, plus the CLI stub.
- Count triggers as four hardcoded paths including ingest_hook.
- Move PreToolUse hook merge out of runner boot into hooks.py.
- Correct the remaining DIRTY INTERFACE.md contracts to match code.
