+++ docs/interfaces/evals/INTERFACE.md
@@ -38,6 +38,8 @@ Result side — `LangfuseEvalRecorder.record()` (`apps/worker/src/curie_worker/e
 
 Model dimension (issue #255, BYO-model epic #24) — additive, same convention as `version:`/`suite:`. A run carries the model id it ran under on `EvalRunResult.model`, which the recorder emits as a `model:<name>` tag plus a `metadata.model` field (and a per-case `metadata.cost_usd` when available). `build_matrix` reads them back (`_model_of`/`_cost_of`) into `EvalCell.model` and a per-model `EvalModelSummary` rollup (pass-rate + summed cost) on `EvalMatrix.model_summaries`, so a suite is sliceable by model for pass-rate/cost comparison. Runs with no resolved model (the `target_url` shortcut) record no `model:` tag and fall into the unlabelled column. The console renders the rollup: the wired evals view shows a "By model" pass-rate and cost table whenever the matrix carries `model_summaries` (`apps/ui/src/views/wired/WiredEvals.tsx`). `cost_usd` is real: the runner prices the token counts the ACI `final` event carries against a local per-model table (`apps/worker/src/curie_worker/eval/pricing.py::cost_usd`, issue #390) and stamps the result on each case row (`apps/worker/src/curie_worker/eval/runner.py::EvalRunner`). That table is deny-by-default and list-price-only, so an unpriced model id, a turn that reported no usage, and the fake tier (no real spend) all record `None` and drop out of the cost rollup instead of counting as free. Pricing deliberately sits on the consumer side rather than on the wire: the protocol carries raw usage, what a token costs is the recorder/matrix's concern.
 
+Matrix cell extras — `EvalCell` (`apps/api/src/curie_api/schemas.py::EvalCell`) also carries `detail` (the scorer's optional explanation, distinct from a trace `error`), `stream_id` (the eval-stream entry that produced the cell), `scorer` (`grader` or `trajectory`), and `case_count`. The recorder writes those as trace metadata (`detail`, `stream_id`, `scorer`, `case_count`; `apps/worker/src/curie_worker/eval/recorder.py::LangfuseEvalRecorder._trace_event`) and `build_matrix` reads them back onto the cell. `EvalMatrix.model_version_summaries` (`apps/api/src/curie_api/schemas.py::EvalMatrix`) slices the model rollup per `(version, model)` so a caller can scope completion to one triggered sha rather than the blended window.
+
 Prompt-cache observability (issue #255) — `_GenerationSpan.record_usage` (`runner/src/curie_runner/otel.py::_GenerationSpan.record_usage`) stamps `cache_read_input_tokens`/`cache_creation_input_tokens` on the generation span, so a warm thread's cache reads surface on the trace — the signal a translating gateway silently broke caching.
 
 ## Implementations today
