+++ docs/interfaces/relational-db/INTERFACE.md
@@ -40,7 +40,7 @@ PostgreSQL 16:
 
 - **DSN + schema** (`apps/api/src/curie_api/db.py::SCHEMA`, `apps/api/src/curie_api/db.py::create_engine`): `SCHEMA = get_settings().db_schema` (default `"curie"`, `apps/api/src/curie_api/config.py::Settings`); the engine is built from `database_url` (env `DATABASE_URL`) via `create_async_engine(..., pool_pre_ping=True)`.
 - **Schema-scoped metadata** (`apps/api/src/curie_api/db.py::Base`): `Base.metadata = MetaData(schema=SCHEMA)` — every table is qualified into the `curie` schema.
-- **Models** (`apps/api/src/curie_api/models.py`): `apps/api/src/curie_api/models.py::Agent` (table `agents`), `apps/api/src/curie_api/models.py::AgentVersion` (table `agent_versions`), `apps/api/src/curie_api/models.py::Deployment` (table `deployments`), `apps/api/src/curie_api/models.py::Approval` (table `approvals`), `apps/api/src/curie_api/models.py::ApprovalAuditEntry` (table `approval_audit_entries`), `apps/api/src/curie_api/models.py::WorkflowStateEntry` (table `workflow_state_entries`), and `apps/api/src/curie_api/models.py::ConsoleSession` (table `console_sessions`), plus the `apps/api/src/curie_api/models.py::Environment` and `apps/api/src/curie_api/models.py::ApprovalStatus` StrEnums. The authoritative list is the set of `Base` subclasses in that module; read it there rather than trusting a count here.
+- **Models** (`apps/api/src/curie_api/models.py`): every `Base` subclass in that module is `apps/api/src/curie_api/models.py::Agent` (table `agents`), `apps/api/src/curie_api/models.py::AgentChannel` (table `agent_channels`), `apps/api/src/curie_api/models.py::AgentVersion` (table `agent_versions`), `apps/api/src/curie_api/models.py::Deployment` (table `deployments`), `apps/api/src/curie_api/models.py::ThreadWorkspace` (table `thread_workspaces`), `apps/api/src/curie_api/models.py::Approval` (table `approvals`), `apps/api/src/curie_api/models.py::Publication` (table `publications`), `apps/api/src/curie_api/models.py::CredentialRedemptionAuditEntry` (table `credential_redemption_audit_entries`), `apps/api/src/curie_api/models.py::ApprovalAuditEntry` (table `approval_audit_entries`), `apps/api/src/curie_api/models.py::AgentAction` (table `agent_actions`), `apps/api/src/curie_api/models.py::ActionAuditEntry` (table `action_audit_entries`), `apps/api/src/curie_api/models.py::WorkflowStateEntry` (table `workflow_state_entries`), and `apps/api/src/curie_api/models.py::ConsoleSession` (table `console_sessions`), plus the `apps/api/src/curie_api/models.py::Environment` and `apps/api/src/curie_api/models.py::ApprovalStatus` StrEnums. The authoritative list remains the set of `Base` subclasses in that module.
 - **Two readers, one DSN** (`apps/api/src/curie_api/db.py::create_engine`, `apps/worker/src/curie_worker/run.py::build`): the API and the worker each build their own `create_async_engine(...)` from their own `database_url` setting (`apps/api/src/curie_api/config.py::Settings`, `apps/worker/src/curie_worker/config.py::WorkerConfig`), both read from `DATABASE_URL`, so a swap repoints both services. The worker never imports the API's models: it reads the schema read-only through hand-written SQL (`apps/worker/src/curie_worker/binding.py::_RESOLVE_SQL`, `apps/worker/src/curie_worker/connector_loop.py::_TARGETS_SQL`), which makes table and column names a second, ORM-independent coupling a conforming DB must honor.
 - **Migrations**: the target DB must apply the **whole Alembic chain in `apps/api/alembic/versions/`**, in revision order, ending at `alembic heads`. The chain grows with the product, so it is deliberately not enumerated here: `ls apps/api/alembic/versions/` is the list, and `alembic heads` is the tip a conforming DB must reach. A single head is the invariant — a fork means two branches each added a migration (rebase and merge the heads before swapping anything).
