+++ docs/interfaces/sealed-credential/INTERFACE.md
@@ -71,11 +71,14 @@ below verbatim; there is nothing else to implement against.
   (`apps/worker/src/curie_worker/sealing.py::open_all`) is all-or-nothing across
   one connector's values: a connector brought up with three of its four
   credentials is the same silently broken pod as one brought up with none.
-- **Failure is fatal by design.** `SealedSecretError`
-  (`apps/worker/src/curie_worker/sealing.py::SealedSecretError`) stops the agent
-  being reconciled, because the alternative — deploy the connector without the
-  credential — produces a pod that starts, passes its health check, and returns
-  401 on every call: healthy in `kubectl get pods` and broken for every user.
+- **Failure is fatal by design, and currently unwired.** `SealedSecretError`
+  (`apps/worker/src/curie_worker/sealing.py::SealedSecretError`) is intended to
+  stop the agent being reconciled, because the alternative — deploy the connector
+  without the credential — produces a pod that starts, passes its health check,
+  and returns 401 on every call: healthy in `kubectl get pods` and broken for
+  every user. Worker production `src` never imports `sealing`; `open_all` /
+  `SealedSecretError` exist only in `sealing.py` and its tests, so the opener
+  does not currently stop reconcile.
 - **Carrier.** `ConnectorSpec.sealed_secrets`
   (`packages/plugin-format/src/plugin_format/connectors.py::ConnectorSpec`), a
   mapping keyed by env var name so it reads like `env` and cannot express two
@@ -143,8 +146,11 @@ file would be misleading without saying so:
   key-usage limits and non-exportable keys; anything that can exec into the
   worker pod reads the raw key, and the CLI deliberately reads the private key
   back out of the cluster to derive the public half. ADR-0094 points at the
-  chart's `existingSecret` idiom as the mitigation, and there is no
-  `sealing.existingSecret` value.
+  chart's `existingSecret` idiom as the mitigation. The shipped values are
+  `sealing.privateKeyExistingSecret` and
+  `sealing.previousPrivateKeyExistingSecret` (`charts/curie/values.yaml`). A
+  plain `curie cluster up` preserves the BYO reference and key without generating
+  an unused chart private key (#1801).
 - **The conformance pin runs one direction and cannot be regenerated by any
   command in the repo.** Rust to Python is fixed by the captured fixture above;
   Python to Rust is exercised only against blobs Rust itself sealed. The fixture's
